+++ src/pages/home/ZoneReport.js
@@ -66,6 +66,7 @@ class ZoneReport extends React.PureComponent {
         <SafeAreaView style={styles.container}>
           {SensorSet1 === 'flour' && SensorSet2 === 'flour' ? null : (
             <View style={styles.areaAir}>
+              <Text style={styles.titleE}>{Strings.getLang('airtemp')}</Text>
               <View style={styles.air}>
                 <FontAwesomeIcon icon={faBacon} color="#00d0ff" size={20} marginRight={10} />
                 <Text style={styles.num}>
@@ -73,7 +74,6 @@ class ZoneReport extends React.PureComponent {
                   °C
                 </Text>
               </View>
-              <Text style={styles.titleE}>{Strings.getLang('airtemp')}</Text>
             </View>
           )}
         </SafeAreaView>
@@ -81,57 +81,53 @@ class ZoneReport extends React.PureComponent {
           {this.stateZ1.isHidden ? (
             <View style={styles.area}>
               <FontAwesomeIcon icon={faMapMarkerAlt} color="#ffb700" size={25} margin={10} />
-              <Text style={styles.titlekwh}>20°C</Text>
+              <Text style={styles.titleE}>{Strings.getLang('zone1')}</Text>
               <Text style={styles.num}>
                 {FaultAlarm === 0 ? t11 : FaultAlarm === 4 ? t11 : FaultAlarm === 8 ? t11 : '--'}
                 °C
               </Text>
               <View>
-                <Text style={styles.title}>{Strings.getLang('now_temp')}</Text>
-                <Text style={styles.titleE}>{Strings.getLang('zone1')}</Text>
+                <Text style={styles.titleE}>{Strings.getLang('now_temp')}</Text>
               </View>
             </View>
           ) : null}
           <View style={styles.area}>
             <FontAwesomeIcon icon={faLightbulb} color="#ffb700" size={25} margin={10} />
-            <Text style={styles.titlekwh}>{Strings.getLang('status')}</Text>
+            <Text style={styles.titleE}>{Strings.getLang('zone1')}</Text>
             <Text style={styles.num}>
               {Relay1flag === true ? Strings.getLang('on') : Strings.getLang('off')}
             </Text>
             <View>
-              <Text style={styles.titlekwh}>
+              <Text style={styles.titleE}>
                 {RelayPower1} 
                 {Strings.getLang('kwh')}
               </Text>
-              <Text style={styles.titleE}>{Strings.getLang('zone1')}</Text>
             </View>
           </View>
           {this.stateZ2.isHidden ? (
             <View style={styles.area}>
               <FontAwesomeIcon icon={faMapMarkerAlt} color="#ff7300" size={25} margin={10} />
-              <Text style={styles.titlekwh}>30°C</Text>
+              <Text style={styles.titleE}>{Strings.getLang('zone2')}</Text>
               <Text style={styles.num}>
                 {FaultAlarm === 0 ? t22 : FaultAlarm === 1 ? t22 : FaultAlarm === 2 ? t22 : '--'}
                 °C
               </Text>
               <View>
-                <Text style={styles.title}>{Strings.getLang('now_temp')}</Text>
-                <Text style={styles.titleE}>{Strings.getLang('zone2')}</Text>
+                <Text style={styles.titleE}>{Strings.getLang('now_temp')}</Text>
               </View>
             </View>
           ) : null}
           <View style={styles.area}>
             <FontAwesomeIcon icon={faLightbulb} color="#ff7300" size={25} margin={10} />
-            <Text style={styles.titlekwh}>{Strings.getLang('status')}</Text>
+            <Text style={styles.titleE}>{Strings.getLang('zone2')}</Text>
             <Text style={styles.num}>
               {Relay2flag === true ? Strings.getLang('on') : Strings.getLang('off')}
             </Text>
             <View>
-              <Text style={styles.titlekwh}>
+              <Text style={styles.titleE}>
                 {RelayPower2} 
                 {Strings.getLang('kwh')}
               </Text>
-              <Text style={styles.titleE}>{Strings.getLang('zone2')}</Text>
             </View>
           </View>
         </SafeAreaView>